Rust sdk eventlog ima support (#67)
* docs(contributor): contrib-readme-action has updated readme

* this commit includes:
- enable IMA event log retrieving and parsing
- add intermediate format and function for eventlog handling both TCG
  fornat and canonical format

* update accoridng to review comments

* update code format

* bug fix for ima log parsing

---------

+++ common/rust/cctrusted_base/src/tdx/quote.rs
@@ -29,12 +29,12 @@ pub struct qgs_msg_get_quote_req {
 
 #[repr(C)]
 pub struct tdx_quote_hdr {
-    pub version: u64,                       // Quote version, filled by TD
-    pub status: u64,                        // Status code of Quote request, filled by VMM
-    pub in_len: u32,                        // Length of TDREPORT, filled by TD
-    pub out_len: u32,                       // Length of Quote, filled by VMM
+    pub version: u64,               // Quote version, filled by TD
+    pub status: u64,                // Status code of Quote request, filled by VMM
+    pub in_len: u32,                // Length of TDREPORT, filled by TD
+    pub out_len: u32,               // Length of Quote, filled by VMM
     pub data_len_be_bytes: [u8; 4], // big-endian 4 bytes indicate the size of data following
-    pub data: [u8; TDX_QUOTE_LEN as usize], // Actual Quote data or TDREPORT on input
+    pub data: [u8; TDX_QUOTE_LEN],  // Actual Quote data or TDREPORT on input
 }
 
 #[repr(C)]
